Wrap flurfunk message body in CDATA
The flurfunk-server will not accept invalid XML, sent by the camelbot if any mails have
subjects that contain German characters (ü is represented as &uuml; - which breaks the
XML).

It should be possible to do this properly though, passing in the same XML as flurfunk-web
does. But just need a quick-fix first.